requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-generativeai==0.8.5
openai==1.3.0
python-dotenv==1.0.0
gunicorn==21.2.0
waitress==2.1.2
